bot_coleta_skincare/requirements.txt
botcity-framework-web
botcity-maestro-sdk
python-dotenv
beautifulsoup4
lxml
webdriver-manager
selenium==4.10.0
